AO3-5752 Allow pinch hitter to look at shared assignment
Previously, the giver (AKA offering user) was preferentially used to determine if the current user was allowed to look at an assignment and to default on it.

This commit adds a clever workaround (== horrible hack) to check for both a giver and a write-in giver, which can happen if both were assigned at "potential matches" stage.