AO3-5748 Show offering byline selectively
In the infrequent case of challenge assignments with two assignees (a giver and a write-in giver AKA pinch hitter), we were previously showing the giver's name, regardless who was looking at the assignment.

Now we will be showing both of them if the user looking is a maintainer, or to each individual their own if the pinch hitter or the giver are looking.