AO3-5901 Introduce a profile presenter
The view was a bit confusing, pulling information from various parts of the user entity, the profile, and the preferences.

In extracting it to a presenter, I realised that if the preference was set to show it, we were showing the birthday section even if the field itself was empty, so I changed that to not show the section.

Refactor some specs to remove duplication.

Update ignore file for Brakeman to ignore the newly introduced false positive.