AO3-TBD Rescue specific error (rubocop)
Rubocop prefers explicitly rescuing `StandardError` rather than an empty / unspecified `rescue`.

Even better would be to rescue a specific error, such as `ActiveRecord::RecordNotDestroyed`, but we don't have the context for why this rescue was needed originally. Given that there is a test for it, it's reasonable to assume that it's something that we expect to occur and we don't want it to block the user's interaction with the site.