AO3-TBD Use standard method when denying access
Only the owner of the request claims is allowed to access them via the user claims page.

Previously, we were using a custom error message and redirect.

Use the standard message that we use in most places in the collection/challenges space - we don't need to give detailed information when a user is trying to access data they don't have access to.